PPO value normalisation
